packages/api/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
mangum==0.19.0
pytest==8.3.5
pydantic
boto3>=1.37.0
boto3-stubs[essential,s3,dynamodb,bedrock-runtime]
strands-agents==0.1.8
strands-agents-tools==0.1.6 
python-dotenv==1.0.0
langchain==0.3.21
langchain-aws==0.2.17
Pillow>=11.2.1
mypy
ruff
pandas
shortuuid
Pillow
tavily-python